requirements.txt
cherrypy
storm
pandas
spacy
xlrd
python-dotenv